Handle config changes for uiMode, mcc, and mnc in car Launcher.
This prevents the activity view from restarting maps whenever the vehicle light sensor changes. One side effect is that the playback fragment does not change to the night theme. This will take more in-depth work to correct; however, since the current fragment will be replaced with a UI update, it makes sense to address at that time. This change is intended to reduce annoyance for current drivers.

Bug: 111096944

Test: build, deploy, toggle night mode property and verify activity doesn't restart and maps reflects day/night mode colors
Change-Id: Iea43165b3ce0590c511052ad1d0325aebcd64e99