DO NOT MERGE Specify display for activities launched from car launcher.
When an activity with an existing task is launched from the launcher and no display is specified, the display for the task is preferred. This creates issues if the task is running in a virtual display (ActivityView).  We want to launch into the main display / the same display as the launcher's context.

Fixes: 131711066

Test: before/after manual - launch Maps activity from launcher while it is hosted in ActivityView, verify launches current task on main display.
Change-Id: Iaa759e3b3c9ca1d7fb8e231cd3bb36c2138ce5fb