Fix AppGrid long click NPE in Launcher
Added a null check to AppLauncherUtils when media services are being processed to check that the intent used to launch the app on a long click is not null. If the intent, returned from PackageManager.getLaunchIntentForPackage is null, the app is launched usi the same intent as the short click.

Added a null check in AppItemViewHolder when setting the long click listener to properly set a no-op for non-media services that are not long clickable.

Bug: 161741917
Test: manual
Change-Id: I4a28068c0c6a9f8c46b1db5f6e0d84ec62037c1e