Fix color of media control button
As part of launcher UX polishing, fixing the color of the media control
button.

Also found a bug with the mediaViewModel not being set properly in the
audio presenter. The presenter should cache the mediaViewModel from its
set models, rather than creating a new one. The model is cached to be
used as the default source of content for the audio card.

Bug: 168939484
Test: manual - built on flame_car
Change-Id: I5e474f53fbac6c21b948c42e8cda92ca91b69257