Increase padding in app grid
Increase padding and reducing margin so that the rotary focus highlight
doesn't abut the app icon. Current padding is 8dp, the same as the
default rotary focus highlight stroke width, so the rotary focus
highlight abuts the app icon. Current margin is 16dp. This CL increases
the padding to 16dp and decreases the margin to 8dp so there's 8dp
between the rotary focus highlight and the app icon.

Test: manual
Bug: 174709055
Change-Id: I04598b74fdb6e64d0fc4000650f3ed06ffb24ddb