Support car reference design in launcher.
With this change launcher will display 2 DisplayAreaGroups(DAG) on the screen. Both the DAG will take half of the screen. The above DAG will display the Maps always and the lower DAG2 will launch rest of the applications in it.
Things that will be done in the follow CL:
1. Remove the maps button from nav bar.
2. Add the ability to show/hide DAG2
3. Add the animations to DAG2

Bug: 186124949
Test: Manual
Change-Id: Ic3c158c5e1b6d17098c3c9607a837230ca352b8d
Merged-In: Ic3c158c5e1b6d17098c3c9607a837230ca352b8d