Add animations to foreground DA hosting default apps
Currently the DA is initialized twice, once with CarLauncher and again with AppGrid.
The animations can be properly seen by clicking the grid view (opens the DA hosting default apps) and then clicking it again (closes the DA).
The work that still needs to be done is:
1. Introduce another DAG to always host CarLauncher with a controlbar.
2. DA's should be registered once.
3. Control all the DAG's from a single activity (register/unregister)

Bug: 186124949
Test: Manual
Change-Id: I285801a006a2a45b20c407cc9a5a50740d0c87e6
Merged-In: I285801a006a2a45b20c407cc9a5a50740d0c87e6