Open correct task in desired DA launched via intents
Currently with these changes if an application is already in the foreground and is launched again the "onTaskInfoChanged" is not invoked. And hence the foreground DA will not show up. Also, when we ask the assistant to give direction currently the maps goto the foreground DA as "onTaskInfoChanged" is not invoked and we cannot reparent it to the correct DA. Once the issue of receinving the callback when activity is in front is resolved the above mentioned problem should be fixed with a followup cl.

Bug: 189109786
Test: Manual
Change-Id: I1b940360d9c864a6a6bc0fd5eb119dfbe4fd1ae4
Merged-In: I1b940360d9c864a6a6bc0fd5eb119dfbe4fd1ae4