Remove unnecessary DA updates onTaskDescriptionChanged
The primary purpose of the listener was to listen for `onActivityRestartAttempt`. When a new task is created, the listener is being called multiple times - once for onTaskCreated and then multiple onTaskDescriptionChanged updates, the latter of which we do not need to listen for. We do not need to listen for onTaskDescriptionChanged to update DA visibility in any other cases.

Bug: 203238640
Test: manual
Change-Id: Ia4b6c486881d2ad5a81b8f0fb057f1f134bc12f5