Release TaskOrganizer when user-switching.
Fixes the issue which shows the blank screen after user-switching.
The issue is caused by when user-switching, the CarLauncher of the
previous user is still responsible as the TaskOrganizer and which
is killed by ATM after some timeout.
Tried to release TaskOrganizer in onDestroy callback, but it wasn't
called, so tries to release it when user-switching.

Bug: 202520182
Test: Check if WelcomeActivity is maintained after user-swtiching.
Test: Check if no regression in gcar_ui_portrait.
Change-Id: I54a62df3786c00ded88095dff7218977c06607e8