Add support switching between full and limited map view in launcher
Introduce and integrate a configuration flag to control whether the
fullscreen or limited map should be embedded into the home screen of the
launcher. The flag can be overridden using static RRO to control the
default behavior per target. Dynamic RROs can be used to configer the
behavior at runtime. Note: the RRO targets are to be added separately.

Bug: 206339678
Bug: 209698294
Test: manual - build and run, validate the default behavior
Test: manual - build and run, apply dynamic RRO and validate the
alternative behavior

Change-Id: I4cf3492ffd9668cb3ad8dc092c7e760dafe71b3a