Fix mute behaviour on the call card
Listen to changes in call audio state in the Launcher Call card and
respond by syncing the mute button's selected state on Launcher
call card.

The Launcher Call card needs to listen to changes in the audio state
that arise from outside of the Launcher, example the Car Dialer app. If
the mute button is toggled in the Dialer app, the Launcher should also
show similar state of mute button.

Added onCallAudioStateChanged to InCallServiceImpl and InCarModel to
listen to mute state changes. According to these changes, update the
selected state of the mute button on Launcher Call card when mute button
is out of sync with global mute state. Added logic to respond to changes
in Drawable icon's selected state in HomeCardFragment.

Bug: 207083521
Test: Manual - tested on hardware, toggled mute button during an ongoing
call on Launcher card (updates Dialer mute button) and
in Dialer app (updates Launcher card mute button). Added unit tests.

Change-Id: Ib7d91675e547cc5adf1da6c416d4917a2024b99a