Retain old state of media widget for non audio use cases
When media source changes to an app that doesn't support media browser session or is a Video App we do not want to update the media widget for the content. Currently we clear out the widget showing
an empty black space. This change will stop clearing the widget state
and let it continue to show the controls for the app that played the media previously.

Bug: 238216936
Fix: 238216936
Test: Manually tested https://screenshot.googleplex.com/8VpYB6LXoLTz7Cu https://photos.app.goo.gl/JBUUtn8dnbeYt3sB7
Change-Id: If879c766bcfc93d481b9b91642374f79455c720f
(cherry picked from commit 7d4f341597e0de0e02402c8516ae9455768902d1)